requirements.txt
google-cloud-bigquery==2.32.0
google-cloud-storage==2.1.0
pandas==1.4.1
pyarrow == 6.0.1
requests==2.27.1